pyproject.toml
[project]
name = "mate"
version = "0.1.0"
description = "Mate OS companion with live captions and automation"
authors = [
    {name = "User",email = "user@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pyside6 (>=6.10.0,<7.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "keyboard (>=0.13.5,<0.14.0)",
    "pywin32 (>=311,<312)",
    "pyperclip (>=1.11.0,<2.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "typer (>=0.20.0,<0.21.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "pyee (>=13.0.0,<14.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "portalocker (>=3.2.0,<4.0.0)",
    "psutil (>=7.1.3,<8.0.0)",
    "pycaw (>=20251023,<20251024)",
    "sounddevice (>=0.5.3,<0.6.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "webrtcvad (>=2.0.10,<3.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)"
]

[project.scripts]
mate = "mate.main:main"
mate-cli = "mate.cli:app"

[tool.poetry]
packages = [{include = "mate", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-qt (>=4.5.0,<5.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "ruff (>=0.14.5,<0.15.0)",
    "types-pywin32 (>=311.0.0.20251008,<312.0.0.0)",
    "coverage (>=7.11.3,<8.0.0)",
    "faker (>=38.0.0,<39.0.0)",
    "httpx[http2] (>=0.28.1,<0.29.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)"
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]
ignore = ["N802", "N803", "N806"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.coverage.run]
branch = true
source = ["mate"]

[tool.coverage.report]
show_missing = true
skip_covered = true
